+++ src/Pim/Bundle/TransformBundle/Cache/DoctrineCache.php
@@ -5,34 +5,27 @@ namespace Pim\Bundle\TransformBundle\Cache;
 use Doctrine\Common\DataFixtures\ReferenceRepository;
 use Doctrine\Common\Persistence\ManagerRegistry;
 use Pim\Bundle\CatalogBundle\Model\ReferableInterface;
+use Pim\Bundle\CatalogBundle\Entity\Repository\ReferableEntityRepositoryInterface;
 
 /**
- * Caches entities for import
+ * Caches doctrine persisted objects
  *
  * @author    Antoine Guigan <antoine@akeneo.com>
  * @copyright 2013 Akeneo SAS (http://www.akeneo.com)
  * @license   http://opensource.org/licenses/osl-3.0.php  Open Software License (OSL 3.0)
  */
-class EntityCache
+class DoctrineCache
 {
-    /**
-     * @var RegistryInterface
-     */
+    /** @var ManagerRegistry */
     protected $doctrine;
 
-    /**
-     * @var ReferenceRepository
-     */
+    /** @var ReferenceRepository */
     protected $referenceRepository;
 
-    /**
-     * @var array
-     */
+    /** @var array */
     protected $cache = array();
 
     /**
-     * Constructor
-     *
      * @param ManagerRegistry $doctrine
      */
     public function __construct(ManagerRegistry $doctrine)
@@ -51,7 +44,7 @@ class EntityCache
     }
 
     /**
-     * Returns an entity by code
+     * Returns an object by code
      *
      * @param string $class
      * @param string $code
@@ -60,11 +53,8 @@ class EntityCache
      */
     public function find($class, $code)
     {
-        if (!isset($this->cache[$class])) {
-            $this->cache[$class] = array();
-        }
-        if (!array_key_exists($code, $this->cache[$class])) {
-            $this->cache[$class][$code] = $this->getEntity($class, $code);
+        if (!isset($this->cache[$class][$code])) {
+            $this->cache[$class][$code] = $this->findObject($class, $code);
         }
 
         return $this->cache[$class][$code];
@@ -90,28 +80,29 @@ class EntityCache
      */
     public function clear()
     {
-        foreach (array_keys($this->cache) as $class) {
-            $this->cache[$class] = array();
-        }
+        $this->cache = array();
     }
 
     /**
-     * Returns an entity from the manager
+     * Returns an object from the manager
      *
      * @param string $class
      * @param string $code
      *
      * @return object
      */
-    protected function getEntity($class, $code)
+    protected function findObject($class, $code)
     {
         $reference = $class . '.' . $code;
         if ($this->referenceRepository && $this->referenceRepository->hasReference($reference)) {
             return $this->referenceRepository->getReference($reference);
         } else {
-            return $this->doctrine
-                    ->getRepository($class)
-                    ->findByReference($code);
+            $repository = $this->doctrine->getManagerForClass($class)->getRepository($class);
+            if (!$repository instanceof ReferableEntityRepositoryInterface) {
+                throw new \Exception(sprintf('Repository "%s" of class "%s" is not referable', get_class($repository), $class));
+            }
+
+            return $repository->findByReference($code);
         }
     }
 }